Move loadBufferFromResources() to resource.h
CUDA depth processor will also use this function.
