+++ examples/protonect/include/libfreenect2/resource.h
@@ -33,6 +33,7 @@ namespace libfreenect2
 {
 
 bool loadResource(const std::string &name, unsigned char const**data, size_t *length);
+bool loadBufferFromResources(const std::string &filename, unsigned char *buffer, const size_t n);
 
 } /* namespace libfreenect2 */
 #endif /* RESOURCE_H_ */
