+++ examples/protonect/src/resource.cpp
@@ -25,6 +25,9 @@
  */
 
 #include <libfreenect2/resource.h>
+#include <string>
+#include <cstring>
+#include <iostream>
 
 namespace libfreenect2
 {
@@ -57,6 +60,27 @@ bool loadResource(const std::string &name, unsigned char const**data, size_t *le
     }
   }
   return result;
+}
+
+bool loadBufferFromResources(const std::string &filename, unsigned char *buffer, const size_t n)
+{
+  size_t length = 0;
+  const unsigned char *data = NULL;
+
+  if(!loadResource(filename, &data, &length))
+  {
+    std::cerr << "loadBufferFromResources: failed to load resource: " << filename << std::endl;
+    return false;
+  }
+
+  if(length != n)
+  {
+    std::cerr << "loadBufferFromResources: wrong size of resource: " << filename << std::endl;
+    return false;
+  }
+
+  memcpy(buffer, data, length);
+  return true;
 }
 
 } /* namespace libfreenect2 */
